Fixing Method references
* Updating the Method item references in the CUI Shortcut items
* Minor Method fixes to support 12.0

The Method items that served as the on_click_handler's of the CUI Shortcut items were being referenced by ID. However, this ID was not consistent with the ID of the actual Method items also being imported. This was likely due to versioning of the Method before the package was exported, so I have updated the references to no longer use the ID and instead just do a query to get the Methods by their names.